requirements.txt
langchain
langchain-community
langchain-groq
groq
streamlit
pandas 
numpy
gradio
googlesearch-python
yfinance
plotly